requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
pydantic[email]
python-multipart
aiofiles
python-dotenv
bcrypt==4.0.1
python-jose[cryptography]
passlib[bcrypt]
httpx
pytest
pytest-asyncio
slowapi
cachetools
email-validator